model-registry/requirements.txt
Flask==3.0.3
pandas==2.2.1
protobuf==5.26.1
pymongo==4.6.3
Requests==2.31.0
SQLAlchemy==2.0.29
Werkzeug==3.0.2
python-dotenv==1.0.1
prometheus-client==0.20.0
prometheus-flask-exporter==0.23.0
flasgger==0.9.7.1
requests-toolbelt==1.0.0
#flask-cors
#flask_httpauth
#gunicorn
# python 3.12.2